+++ CHANGELOG.md
@@ -8,8 +8,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [8.56.0] - 2025-08-26
 
 ### Changed
-
 - Update parent from 5.12.1 to 5.13.0
+  
+- update jeap-spring-boot-roles-anywhere-starter from 1.2.0 to 1.3.0
 
 ## [8.55.0] - 2025-08-22